Virtual Fit v1.0
Added the first complete version for Virtual Fit. In order to run this website on your own device, You need to place the folder in your XAMPP->htdocs folder. For the PHP mailer to work, You need to input your own credentials for the gmail app such that the email gets delivered. The website works fine without this and it is not necessary to setup the email credentials. If you want to change the credentials, Open the workspace in VS Code, Press ctrl+shift+F and search for "arhamahmed8699@gmail.com" & "password" and replace it with your email and the generated password for your app. (You can get your generated password at https://myaccount.google.com/apppasswords)